2004-12-30  Bob Doan  <bdoan@sicompos.com> 	* libsrc/free.c: legends, x axis labels try to fit better, minor cleanups 	* libsrc/graphing.c: likewise 	* libsrc/parsexml.c: likewise 	* libsrc/pcode.h: likewise 	* libsrc/pdf.c: likewise 	* libsrc/resolution.c: likewise 	* libsrc/rlib.h: likewise 	* src/examples/php/graph.php: likewise 	* src/examples/php/graph.xml: likewise

+++ libsrc/pcode.h
@@ -170,7 +170,7 @@ gint format_number(gchar *dest, gint max, const gchar *moneyformat, gint64 x);
 int execute_pcode(rlib *r, struct rlib_pcode *code, struct rlib_value_stack *vs, struct rlib_value *this_field_value);
 struct rlib_value * rlib_value_stack_pop(struct rlib_value_stack *vs);
 int rlib_value_stack_push(struct rlib_value_stack *vs, struct rlib_value *value);
-struct rlib_value * rlib_value_new_number(struct rlib_value *rval, long long value);
+struct rlib_value * rlib_value_new_number(struct rlib_value *rval, gint64 value);
 struct rlib_value * rlib_value_new_string(struct rlib_value *rval, char *value);
 struct rlib_value * rlib_value_new_date(struct rlib_value *rval, struct rlib_datetime *date);
 void rlib_pcode_dump(struct rlib_pcode *p, int offset);
